examples/python/executor/agentkit/requirements.txt
# Core gRPC dependencies
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0

# LLM Providers
openai>=1.12.0
anthropic>=0.18.0

# HTTP requests (for n8n integration)
requests>=2.31.0

# LangChain (for langchain-agent.py)
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-core>=0.1.0
pydantic>=2.0.0

# CrewAI (for crewai-multi-agent.py)
crewai>=0.22.0
crewai-tools>=0.1.0

# AutoGen (for autogen-collaborative.py)
pyautogen>=0.2.0

# Utilities
python-dotenv>=1.0.0
